Extract MMapReader to examples, leaving only SimpleMappedReader in base. Move all deps to examples. JVector proper is now at 0 non-test deps.
