+++ jvector-base/src/main/java/com/github/jbellis/jvector/disk/SimpleMappedReader.java
@@ -28,8 +28,8 @@ import java.util.logging.Logger;
 
 /**
  * DO NOT use this for anything you care about.
- * Lies about implementing interfaces. Bare minimum I/O to run Bench/SiftSmall
- * against disk in reasonable time. Does not handle files above 2 GB.
+ * Lies about implementing interfaces. Bare minimum I/O to run against disk in reasonable time.
+ * Does not handle files above 2 GB.
  */
 public class SimpleMappedReader implements RandomAccessReader {
     private static final Logger LOG = Logger.getLogger(SimpleMappedReader.class.getName());
