+++ jvector-examples/pom.xml
@@ -30,6 +30,16 @@
             <artifactId>jvector-base</artifactId>
             <version>${jvector.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.jhdf</groupId>
+            <artifactId>jhdf</artifactId>
+            <version>0.6.10</version>
+        </dependency>
+        <dependency>
+            <groupId>com.indeed</groupId>
+            <artifactId>util-mmap</artifactId>
+            <version>1.0.52-3042601</version>
+        </dependency>
     </dependencies>
     <profiles>
         <profile>
