+++ jvector-examples/src/main/java/com/github/jbellis/jvector/example/Bench.java
@@ -19,9 +19,9 @@ package com.github.jbellis.jvector.example;
 import com.github.jbellis.jvector.disk.CachingGraphIndex;
 import com.github.jbellis.jvector.disk.CompressedVectors;
 import com.github.jbellis.jvector.disk.OnDiskGraphIndex;
-import com.github.jbellis.jvector.disk.ReaderSupplierFactory;
 import com.github.jbellis.jvector.example.util.DataSet;
 import com.github.jbellis.jvector.example.util.Hdf5Loader;
+import com.github.jbellis.jvector.example.util.ReaderSupplierFactory;
 import com.github.jbellis.jvector.graph.*;
 import com.github.jbellis.jvector.pq.ProductQuantization;
 import com.github.jbellis.jvector.vector.VectorEncoding;
