+++ jvector-examples/src/main/java/com/github/jbellis/jvector/example/util/MMapReaderSupplier.java
@@ -1,5 +1,7 @@
-package com.github.jbellis.jvector.disk;
+package com.github.jbellis.jvector.example.util;
 
+import com.github.jbellis.jvector.disk.RandomAccessReader;
+import com.github.jbellis.jvector.disk.ReaderSupplier;
 import com.indeed.util.mmap.MMapBuffer;
 
 import java.io.IOException;
