+++ pom.xml
@@ -145,16 +145,6 @@
     </build>
     <!-- Only include dependencies that are needed for the multirelease jar to run -->
     <dependencies>
-        <dependency>
-            <groupId>io.jhdf</groupId>
-            <artifactId>jhdf</artifactId>
-            <version>0.6.10</version>
-        </dependency>
-        <dependency>
-            <groupId>com.indeed</groupId>
-            <artifactId>util-mmap</artifactId>
-            <version>1.0.52-3042601</version>
-        </dependency>
     </dependencies>
     <dependencyManagement>
         <dependencies>
